Module Rtc

module IceServer : sig ... end

module IdentityAssertion : sig ... end

module IceCandidate : sig ... end

module DtlsTransport : sig ... end

module RtpSender : sig ... end

module RtpReceiver : sig ... end

module RtpTransceiver : sig ... end

module SctpTransport : sig ... end

Data Channels

The data channels allows us to represent a network channel which can be used for bidirectional peer-to-peer transfer of data.

module DataChannel : sig ... end

Peer Connections

module PeerConnection : sig ... end