Module Rtc.IceCandidate

type t

A candidate internet connectivity establishment configuration which can be used to establish a PeerConnection.

include Jv.CONV with type t := t

val to_jv : t -> Jv.jv

val of_jv : Jv.jv -> t

Properties

val get_candidate : t -> Jstr.t

get_candidate t returns the transport address in the format described in RFC5245.

val get_component : t -> [ `RTP | `RTCP ]

get_component t identifies the candidate as either for RTP or RTCP.

val get_foundation : t -> Jstr.t

See the docs

val get_ip : t -> Jstr.t

get_ip t is the IP address of the candidate

val get_port : t -> int

get_port t is the integer port number of the candidate

val get_priority : t -> int

get_priority t is the candidates priority

val get_address : t -> Jstr.t

get_address t is the candidate's address

val get_protocol : t -> [ `TCP | `UDP ]

get_protocol t is the candidates protocol

val get_related_port : t -> int option

The port number of the reflexive or relay candidates of this candidate

val get_sdp_mid : t -> Jstr.t option

get_sdp_mid t is the candidate's media stream identification tag

val get_sdp_m_line_index : t -> int option

See docs

val get_tcp_type : t -> [ `Active | `Passive | `So ] option

If a TCP connection this is the type of connection. `Active means the transport will open an outbound connectiion but won't receive incoming connection requests, `Passive is the opposite and `So means the transport will try to open a connection simultaneously with the peer.

Methods

val to_json : t -> Brr.Json.t

to_json t is the JSON representing the configuration of the IceCandidate configuration