Module Rtc.PeerConnection

module Bundle : sig ... end

The bundle policy is part of the RTCConfiguration for peer connections. Bundle is a feature of the Session Description Protocol (SDP) that allows you to specify a media flow to have the same source and destination addresses and ports, along with the same udnerlyign transfer protocol. WebRTC Standards has further information.

module Certificate : sig ... end

module IceTransportPolicy : sig ... end

module RtcpMuxPolicy : sig ... end

type config

RTCConfiuration type

val config : ?bundle_policy:Bundle.t -> ?certificates:Certificate.t array -> ?ice_candidate_pool_size:int -> ?ice_servers:IceServer.t array -> ?ice_transport_policy:IceTransportPolicy.t -> ?peer_identity:Jstr.t -> ?rtcp_mux_policy:RtcpMuxPolicy.t -> unit -> config

Builder for the RTCConfiguration dictionary

type t

An RTCPeerConnection

include Jv.CONV with type t := t

val to_jv : t -> Jv.jv

val of_jv : Jv.jv -> t

val as_target : t -> Brr.Ev.target

as_target t converts t to an event target

val create : ?opts:config -> unit -> t

create ?config () is a new RTC peer connection using the optional config configuration

val can_trickle_ice_candidates : t -> bool

The bool returned indicates whether or not the remote peer can accept trickled ice candidates.

module State : sig ... end

val get_connection_state : t -> State.t

module Sd : sig ... end

val get_current_local_description : t -> Sd.t

current_local_description t returns session description protocol if it can be parsed otherwise and error along with the underlying SDP in string format

val get_current_remote_description : t -> Sd.t

same as current_local_description only describing the remote end of the connection

module IceState : sig ... end

val get_ice_connection_state : t -> IceState.t

ice_connection_state t is the ICE agent state for peer connection t

val get_ice_gathering_state : t -> [ `New | `Gathering | `Complete ]

ice_gathering_state t describes peer connection t's ICE gathering state. `New means the peer connection was just created, `Gathering means the ICE agent is in the process or gathering candidates and `Completed means it has finished.

val get_local_description : t -> Sd.t option

local_description t describes the session for the local end of the connection

val get_peer_identity : t -> IdentityAssertion.t Fut.or_error

peer_identity t returns a promise that resolves to an identity assertion containing the string identifying the remote peer.

val get_pending_local_description : t -> Sd.t

pending_local_description t returns the session description protocol if it can be parsed otherwise and error along with the underlying SDP in string format for a potential future description not the current one

val get_pending_remote_description : t -> Sd.t

same as pending_local_description only describing the remote end of the connection

val get_remote_description : t -> Sd.t

same as local_description only for describing the remote end of the conneciton

val get_sctp : t -> SctpTransport.t option

sctp t returns the SctpTransport describing the SCTP transport over which SCTP data is being sent and received. It will be None if it has not been negotiated

val get_ready_state : t -> DataChannel.State.t

get_ready_state

module SignalingState : sig ... end

val get_signaling_state : t -> SignalingState.t

signaling_state t descibes the state of the signaling process on the local end of the connection while connecting or reconnecting to another peer

Events

module Ev : sig ... end

Methods

Methods are the functions you can call on your peer connection. For the latest documentation check the MDN pages.

val add_ice_candidate : ?candidate:IceCandidate.t -> connection:t -> unit -> unit Fut.or_error

add_ice_candidate ~candidate ~connection () will add the canidate to the connection's remote description. If no candidate is specified then an end-of-candidates indicator is added signaling all remote candidates have been sent.

val add_track : ?stream:Brr_io.Media.Stream.t -> track:Brr_io.Media.Track.t -> connection:t -> unit -> RtpSender.t

add_track ~track ~connection () adds a new track to the set of tracks that will be sent to the peer

val close : t -> unit

Closes the peer connection

type offer

val offer : ?ice_restart:bool -> ?voice_activity_detection:bool -> unit -> offer

val create_offer : ?offer:offer -> t -> Sd.t Fut.or_error

Initiate the creation of an SDP offer

val create_answer : ?answer:offer -> t -> Sd.t Fut.or_error

Creates an SDP answer to an offer received from a remote peer during the offer/answer negotiation

val create_data_channel : ?opts:DataChannel.opts -> label:string -> t -> DataChannel.t

val generate_certificate : Brr_webcrypto.Crypto_algo.t -> Certificate.t Fut.or_error

A static function that creates an X.509 certificate corresponding private key

val get_senders : t -> RtpSender.t array

Returns an array RTP senders, each responsible for sending one track's data

module StatsReport : sig ... end

val get_stats : ?selector:Brr_io.Media.Track.t -> t -> StatsReport.t Fut.or_error

get_stats t will returns the stats for the whole connection

val get_transceivers : t -> RtpTransceiver.t array

get_transceivers t gets all of the peer connection's transceivers in the order they were added. You could stop all of the transceivers associated with a peer connection using:

Array.iter RtpTransceiver.stop (get_transceivers t) 

val remove_track : RtpSender.t -> t -> unit

remove_track sender t will remove sender from the peer connection t, this will stop sending media from the track for the local end of the connection.

val restart_ice : t -> unit

restart_ice t will request that ICE candidate gathering be redone on both ends of the connection. You might want to use this on a state change and the state is Failed.

val set_local_description : Sd.t -> t -> unit Fut.or_error

val set_remote_description : Sd.t -> t -> unit Fut.or_error