Module Rtc.SctpTransport

type t

Stream Control Transmission Protocol description

val as_target : t -> Brr.Ev.target

as_target t converts t to an event target

include Jv.CONV with type t := t

val to_jv : t -> Jv.jv

val of_jv : Jv.jv -> t

val get_max_channels : t -> int

get_max_channels t indicates the maximum number of DataChannels that can be open simultaneously

val get_max_message_size : t -> int

get_max_message_size t indicate in bytes, the maximum size of a message than can be sent over the DataChannel

module State : sig ... end

val state : State.t Brr.Ev.Type.t

val get_state : t -> State.t

get_state t is the state of the SCTP transport

val get_transport : t -> DtlsTransport.t

get_transport t returns the DtlsTransport object used for the transmission and receipt of the packets